requirements.txt
PyYAML
Flask
pytest
streamlit